refactor: details 정리
