+++ src/features/bid/api/index.ts
@@ -1,1 +1,2 @@
+export { cancelBid } from './cancelBid';
 export { postBid } from './postBid';
