+++ src/features/bid/config/index.ts
@@ -1,4 +1,3 @@
 export { BID_CAUTION, EDIT_BID_CAUTION } from './caution';
-export { MAX_BID_COUNT } from './constants';
 export { getBidSchema } from './schema';
 export type { IBidPostData, IBidder, IBidderList } from './type';
