+++ src/features/bid/index.ts
@@ -1,1 +1,2 @@
+export { useCancelBid } from './model/index';
 export { BidForm } from './ui';
