+++ src/features/bid/model/index.ts
@@ -1,1 +1,2 @@
+export { useCancelBid } from './useCancelBid';
 export { usePostBid } from './usePostBid';
