+++ src/features/details/api/convertAuction.ts
@@ -1,0 +1,7 @@
+import { API_END_POINT, httpClient } from '@/shared';
+
+export const convertAuction = async (preAuctionId: number) => {
+  const response = await httpClient.post(`${API_END_POINT.AUCTION}/start`, preAuctionId);
+
+  return response.data;
+};
