+++ src/features/details/api/deletePreAuction.ts
@@ -1,0 +1,6 @@
+import { API_END_POINT, httpClient } from '@/shared';
+
+export const deletePreAuction = async (preAuctionId: number) => {
+  await httpClient.delete(`${API_END_POINT.PRE_AUCTION}/${preAuctionId}`);
+  return;
+};
