+++ src/features/details/index.ts
@@ -1,0 +1,2 @@
+export { useDeletePreAuction, useGetAuctionDetails, useGetPreAuctionDetailsWithSuspense } from './model';
+export { AuctionDetailsMain, PreAuctionDetailsMain } from './ui';
