+++ src/features/details/lib/index.ts
@@ -1,0 +1,1 @@
+export { useEndAuction } from './useEndAuction';
