+++ src/features/details/model/useGetPreAuctionDetailsWithSuspense.ts
@@ -1,0 +1,14 @@
+import { QUERY_KEYS } from '@/shared';
+import { useSuspenseQuery } from '@tanstack/react-query';
+import { getPreAuctionDetails } from '../api';
+
+export const useGetPreAuctionDetailsWithSuspense = (preAuctionId: number) => {
+  const { data: preAuctionDetails } = useSuspenseQuery({
+    queryKey: [QUERY_KEYS.PRE_AUCTION_DETAILS, preAuctionId],
+    queryFn: () => getPreAuctionDetails(preAuctionId),
+  });
+
+  return {
+    preAuctionDetails,
+  };
+};
