+++ src/features/details/ui/index.ts
@@ -1,0 +1,7 @@
+export { AuctionDetailsFooter } from './AuctionDetailsFooter';
+export { AuctionDetailsMain } from './AuctionDetailsMain';
+export { DetailsBasic } from './DetailsBasic';
+export { DetailsOption } from './DetailsOption';
+export { PreAuctionDetailsFooter } from './PreAuctionDetailsFooter';
+export { PreAuctionDetailsMain } from './PreAuctionDetailsMain';
+export { ProgressBar } from './ProgressBar';
