+++ src/shared/constants/bid.ts
