+++ src/shared/constants/index.ts
@@ -1,7 +1,8 @@
 export { API_END_POINT } from './apiEndPoint';
+export { MAX_BID_COUNT } from './bid';
+export { CATEGORIES } from './categories';
 export { CONFIRM_MESSAGE } from './confirmMessage';
 export { EMPTY_MESSAGE } from './emptyMessage';
 export { ERROR_CODE } from './errorCode';
 export { QUERY_KEYS } from './queryKeys';
 export { ROUTES } from './routes';
-export { CATEGORIES } from './categories';
